Refactor check_reply into modular structure
