+++ src/errors.rs
@@ -39,4 +39,3 @@ pub enum Error {
     #[error("configuration error: {0}")]
     Config(String),
 }
-
