+++ src/lib.rs
@@ -1,2 +1,3 @@
-pub mod repository;
+pub mod check_reply;
 pub mod errors;
+pub mod repository;
